requirements.txt
anyio==3.6.2
async-timeout==4.0.2
bidict==0.22.1
click==8.1.3
dnspython==2.3.0
fastapi==0.89.1
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.4
pymongo==4.3.3
python-dotenv==0.21.1
python-engineio==4.3.4
python-socketio==5.7.2
PyYAML==6.0
redis==4.4.2
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
